Split types in UIElements.swift into seperate files.
